Use correct type for wz.

+++ libz80e/include/z80e/core/registers.h
@@ -60,7 +60,7 @@ typedef struct {
 		};
 	};
 	uint8_t I, R;
-	uint8_t WZ;
+	uint16_t WZ;
 } z80registers_t;
 
 typedef enum {
